Don't clone a repo twice during force install
A force install (`newt install -f`) deletes the existing repos and
clones them anew.  If the repo hadn't been installed yet, a force
install executed the following sequence:

1. Clone repo in order to get its `repository.yml` file.
2. Delete repo due to `-f` flag.
3. Clone repo again.

This commit causes steps 2 and 3 to be skipped if the repo didn't
originally exist, as a new copy just got cloned.

+++ newt/install/install.go
@@ -729,8 +729,9 @@ func (inst *Installer) Install(
 	// For a forced install, delete all existing repos.
 	if force {
 		for _, r := range repos {
-			// Don't delete the local project directory!
-			if !r.IsLocal() {
+			// Don't delete the local project directory!  And don't delete a
+			// repo that was just cloned during this invocation of newt.
+			if !r.IsLocal() && !r.IsNewlyCloned() {
 				util.StatusMessage(util.VERBOSITY_DEFAULT,
 					"Removing old copy of \"%s\" (%s)\n", r.Name(), r.Path())
 				os.RemoveAll(r.Path())
